api/requirements.txt
psycopg2-binary==2.8.6
requests == 2.25.1
jsonpath == 0.82
pytest == 6.2.2